backend/requirements.txt
flask
flask-sqlalchemy
flask-admin
gunicorn
pandas
python-decouple
python-dotenv
psycopg2
flask-cors